finally a sane structure. Created db folder
