+++ daemon-rs/diesel.toml
@@ -1,3 +1,3 @@
 
 [print_schema]
-file = "schema.rs"
+file = "db/schema.rs"
