+++ daemon-rs/global.hpp
@@ -2,8 +2,7 @@
 
 #include "constants.hpp"
 #include "crypto/crypto.hpp"
-#include "db.hpp"
-#include "constants.hpp"
+#include "db/db.hpp"
 
 // Global is a thread safe singleton that all threads should have
 // access to. It exists to provide notifications as well as set up the database.
