+++ daemon-rs/main.cc
@@ -9,7 +9,7 @@
 
 #include "asphr/asphr.hpp"
 #include "client_lib/client_lib.hpp"
-#include "db.hpp"
+#include "daemon-rs/db/db.hpp"
 #include "global.hpp"
 // #include "rpc/daemon_rpc.hpp"
 #include "schema/server.grpc.pb.h"
