+++ daemon-rs/transmitter/transmitter.hpp
@@ -5,10 +5,10 @@
 
 #pragma once
 
-#include "../crypto/crypto.hpp"
-#include "../db.hpp"
-#include "../global.hpp"
 #include "asphr/asphr.hpp"
+#include "daemon-rs/crypto/crypto.hpp"
+#include "daemon-rs/db/db.hpp"
+#include "daemon-rs/global.hpp"
 #include "pir/fast_pir/fast_pir_client.hpp"
 #include "schema/daemon.grpc.pb.h"
 #include "schema/server.grpc.pb.h"
